+++ examples/execd/tests/integration_test.rs
@@ -13,12 +13,14 @@ use libcnb_test::{assert_contains, IntegrationTest};
 #[ignore]
 fn basic() {
     IntegrationTest::new("heroku/buildpacks:20", "test-fixtures/empty-app").run_test(|context| {
-        context.prepare_container().start(|container| {
-            let env_stdout = container.shell_exec("env").stdout;
+        context
+            .prepare_container()
+            .start_with_shell_command("env", |container| {
+                let env_stdout = container.logs_wait().stdout;
 
-            assert_contains!(env_stdout, "ROLL_1D6=");
-            assert_contains!(env_stdout, "ROLL_4D6=");
-            assert_contains!(env_stdout, "ROLL_1D20=");
-        });
+                assert_contains!(env_stdout, "ROLL_1D6=");
+                assert_contains!(env_stdout, "ROLL_4D6=");
+                assert_contains!(env_stdout, "ROLL_1D20=");
+            });
     });
 }